Add RoundChange v2 types and functions
