+++ consensus/istanbul/core/handler.go
@@ -34,6 +34,7 @@ func (c *core) Start() error {
 
 	c.current = roundState
 	c.roundChangeSet = newRoundChangeSet(c.current.ValidatorSet())
+	c.roundChangeSetV2 = newRoundChangeSetV2(c.current.ValidatorSet())
 
 	// Reset the Round Change timer for the current round to timeout.
 	// (If we've restored RoundState such that we are in StateWaitingForRoundChange,
